Add resin specific environment configuration
This patch defines all the needed bits on which resinhup relies when swiching
the root partition. The interface with uboot is done through an env file
(resinOS_uEnv.txt) located in the first partition which, if it exists, might
overwrite the default root partition index. In order to overwrite this, the env
file must define `resin_root_part`.

Signed-off-by: Andrei Gherzan <andrei@resin.io>

Upstream-Status: Inappropriate [configuration]

%% original patch: resin-specific-env-configuration.patch